:recycle: timeline related component relocated
